SGAT/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
blas=1.0=mkl
brotlipy=0.7.0=py39h3811e60_1001
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py39hf3d152e_1
cffi=1.14.5=py39he32792d_0
chardet=4.0.0=py39hf3d152e_1
cryptography=3.4.4=py39h95dcef6_0
cudatoolkit=10.2.89=h8f6ccaa_8
decorator=4.4.2=py_0
dgl-cuda10.2=0.6.0=py39_0
gudhi=3.4.1.post1=pypi_0
idna=2.10=pyh9f0ad1d_0
joblib=1.0.1=pypi_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libstdcxx-ng=9.3.0=h6de172a_18
llvm-openmp=11.0.1=h4bd325d_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py39h07f9747_2
mkl_fft=1.2.0=py39h1dff97c_1
mkl_random=1.0.2=py39h63df603_0
ncurses=6.2=h58526e2_4
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_0
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
openssl=1.1.1j=h7f98852_0
pickle5=0.0.11=pypi_0
pip=20.0.2=py_2
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py39hf3d152e_3
python=3.9.2=hffdb5ce_0_cpython
python_abi=3.9=1_cp39
readline=8.0=he28a2e2_2
requests=2.25.1=pyhd3deb0d_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.1=py39h91f5cce_0
setuptools=49.6.0=py39hf3d152e_3
six=1.15.0=pyh9f0ad1d_0
sklearn=0.0=pypi_0
sqlite=3.34.0=h74cdb3f_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=h21135ba_1
torch=1.8.0=pypi_0
typing_extensions=3.7.4.3=py_0
tzdata=2021a=he74cb21_0
urllib3=1.26.3=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
